requirements.txt
geopy==2.3.0
lxml==4.9.2
Pillow==9.5.0
